Update interfaces and fix database precision issues
This commit consolidates the following changes:
- Update four interfaces for address, token and transaction operations
- Migrate amount columns from bigint to decimal to avoid precision problems
- Fix token search API and add fuzzy search functionality
- Return holdersNum as a string in response objects
- Update migration database files
- Revise code issues identified in review

+++ packages/tracker/src/entities/txOutArchive.entity.ts
@@ -25,7 +25,7 @@ export class TxOutArchiveEntity {
   @Index()
   ownerPubKeyHash: string;
 
-  @Column({ name: 'token_amount', type: 'bigint', nullable: true })
+  @Column({ name: 'token_amount', type: 'decimal', nullable: true })
   tokenAmount: bigint;
 
   @Column({ name: 'data' })
